feat: implement rem, div, imulc
