+++ intrinsics/Cargo.toml
@@ -3,7 +3,7 @@ name = "ethnum-intrinsics"
 version = "1.2.0"
 authors = ["Nicholas Rodrigues Lordello <nlordell@gmail.com>"]
 edition = "2021"
-description = "LLVM generated 256-bit integer intrinsics"
+description = "C 256-bit integer intrinsics"
 documentation = "https://docs.rs/ethnum-intrinsics"
 readme = "README.md"
 homepage = "https://github.com/nlordell/ethnum-rs"
