+++ intrinsics/README.md
@@ -1,6 +1,6 @@
-# LLVM Generated Intrinsics
+# C Intrinsics
 
-This crate contains a build script and Rust bindings to LLVM generated
+This crate contains a build script and Rust bindings to C generated
 intrinsics. These intrinsics can optionally be used by `ethnum` crate and aren't
 intended to be used directly. See `ethnum` crate README documentation for more
 information.
