+++ src/intrinsics/native.rs
@@ -4,10 +4,13 @@
 mod add;
 mod ctz;
 mod divmod;
+mod divmod_shared;
 mod mul;
 mod rot;
 mod shl;
 mod shr;
 mod sub;
 
-pub use self::{add::*, ctz::*, divmod::*, mul::*, rot::*, shl::*, shr::*, sub::*};
+pub use self::{
+    add::*, ctz::*, divmod::*, divmod_shared::*, mul::*, rot::*, shl::*, shr::*, sub::*,
+};
